// src/app/interfaces/task.interface.ts
// ○ ID(required)
// ○ Project ID(required)
// ○ User(required)
// ○ Title(required)
// ○ Slug(required, autogenerated)
// ○ Description(optional)
// ○ Due Date(optional)
// ○ Status(Default to ‘To Do’, required, list of[‘To Do’, ’Doing’, ’Done’])

export interface Task {
  id: string,
  project_id: string,
  user: string,
  title: string,
  slug: string,
  description?: string,
  due_date?: string
  status: TaskStatus
}

export enum TaskStatus {
  to_do = 0,
  doing = 1,
  done = 2
}